chore: move backup cmd into dataprotection cmd
