+++ docs/user_docs/cli/cli.md
@@ -25,16 +25,6 @@ Manage alert receiver, include add, list and delete receiver.
 * [kbcli alert list-smtpserver](kbcli_alert_list-smtpserver.md)	 - List alert smtp servers config.
 
 
-## [backup](kbcli_backup.md)
-
-Backup command.
-
-* [kbcli backup create](kbcli_backup_create.md)	 - Create a backup for the cluster.
-* [kbcli backup delete](kbcli_backup_delete.md)	 - Delete a backup.
-* [kbcli backup describe](kbcli_backup_describe.md)	 - Describe a backup
-* [kbcli backup list](kbcli_backup_list.md)	 - List backups.
-
-
 ## [backuprepo](kbcli_backuprepo.md)
 
 BackupRepo command.
@@ -158,6 +148,13 @@ List and open the KubeBlocks dashboards.
 
 * [kbcli dashboard list](kbcli_dashboard_list.md)	 - List all dashboards.
 * [kbcli dashboard open](kbcli_dashboard_open.md)	 - Open one dashboard.
+
+
+## [dataprotection](kbcli_dataprotection.md)
+
+Data protection command.
+
+* [kbcli dataprotection backup](kbcli_dataprotection_backup.md)	 - Backup command.
 
 
 ## [fault](kbcli_fault.md)
