+++ docs/user_docs/cli/kbcli.md
@@ -56,7 +56,6 @@ kbcli [flags]
 
 * [kbcli addon](kbcli_addon.md)	 - Addon command.
 * [kbcli alert](kbcli_alert.md)	 - Manage alert receiver, include add, list and delete receiver.
-* [kbcli backup](kbcli_backup.md)	 - Backup command.
 * [kbcli backuprepo](kbcli_backuprepo.md)	 - BackupRepo command.
 * [kbcli bench](kbcli_bench.md)	 - Run a benchmark.
 * [kbcli builder](kbcli_builder.md)	 - builder command.
@@ -66,6 +65,7 @@ kbcli [flags]
 * [kbcli clusterversion](kbcli_clusterversion.md)	 - ClusterVersion command.
 * [kbcli context](kbcli_context.md)	 - kbcli context allows you to manage cloud context. This command is currently only applicable to cloud, and currently does not support switching the context of the local k8s cluster.
 * [kbcli dashboard](kbcli_dashboard.md)	 - List and open the KubeBlocks dashboards.
+* [kbcli dataprotection](kbcli_dataprotection.md)	 - Data protection command.
 * [kbcli fault](kbcli_fault.md)	 - Inject faults to pod.
 * [kbcli infra](kbcli_infra.md)	 - infra command
 * [kbcli kubeblocks](kbcli_kubeblocks.md)	 - KubeBlocks operation commands.
