+++ docs/user_docs/cli/kbcli_dataprotection_backup_create.md
@@ -1,0 +1,61 @@
+---
+title: kbcli dataprotection backup create
+---
+
+Create a backup for the cluster.
+
+```
+kbcli dataprotection backup create NAME [flags]
+```
+
+### Examples
+
+```
+  # Create a backup for the cluster
+  kbcli backup create mybackup --cluster mycluster
+  
+  # create a snapshot backup
+  kbcli backup create mybackup --cluster mycluster --type snapshot
+  
+  # create a backup with specified policy
+  kbcli backup create mybackup --cluster mycluster --policy mypolicy
+```
+
+### Options
+
+```
+      --cluster string   Cluster name
+  -h, --help             help for create
+      --method string    Backup type (default "snapshot")
+      --policy string    Backup policy name, this flag will be ignored when backup-type is snapshot
+```
+
+### Options inherited from parent commands
+
+```
+      --as string                      Username to impersonate for the operation. User could be a regular user or a service account in a namespace.
+      --as-group stringArray           Group to impersonate for the operation, this flag can be repeated to specify multiple groups.
+      --as-uid string                  UID to impersonate for the operation.
+      --cache-dir string               Default cache directory (default "$HOME/.kube/cache")
+      --certificate-authority string   Path to a cert file for the certificate authority
+      --client-certificate string      Path to a client certificate file for TLS
+      --client-key string              Path to a client key file for TLS
+      --context string                 The name of the kubeconfig context to use
+      --disable-compression            If true, opt-out of response compression for all requests to the server
+      --insecure-skip-tls-verify       If true, the server's certificate will not be checked for validity. This will make your HTTPS connections insecure
+      --kubeconfig string              Path to the kubeconfig file to use for CLI requests.
+      --match-server-version           Require server version to match client version
+  -n, --namespace string               If present, the namespace scope for this CLI request
+      --request-timeout string         The length of time to wait before giving up on a single server request. Non-zero values should contain a corresponding time unit (e.g. 1s, 2m, 3h). A value of zero means don't timeout requests. (default "0")
+  -s, --server string                  The address and port of the Kubernetes API server
+      --tls-server-name string         Server name to use for server certificate validation. If it is not provided, the hostname used to contact the server is used
+      --token string                   Bearer token for authentication to the API server
+      --user string                    The name of the kubeconfig user to use
+```
+
+### SEE ALSO
+
+* [kbcli dataprotection backup](kbcli_dataprotection_backup.md)	 - Backup command.
+
+#### Go Back to [CLI Overview](cli.md) Homepage.
+
