--- internal/cli/cmd/backup/describe.go
@@ -1,59 +1,0 @@
-/*
-Copyright (C) 2022-2023 ApeCloud Co., Ltd
-
-This file is part of KubeBlocks project
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-package backup
-
-import (
-	"github.com/spf13/cobra"
-	"k8s.io/cli-runtime/pkg/genericclioptions"
-	cmdutil "k8s.io/kubectl/pkg/cmd/util"
-	"k8s.io/kubectl/pkg/util/templates"
-
-	"github.com/apecloud/kubeblocks/internal/cli/cmd/cluster"
-	"github.com/apecloud/kubeblocks/internal/cli/printer"
-	"github.com/apecloud/kubeblocks/internal/cli/types"
-	"github.com/apecloud/kubeblocks/internal/cli/util"
-)
-
-var (
-	describeBackupExample = templates.Examples(`
-		# describe a backup
-		kbcli backup describe mybackup
-	`)
-)
-
-func newDescribeCommand(f cmdutil.Factory, streams genericclioptions.IOStreams) *cobra.Command {
-	o := cluster.DescribeBackupOptions{
-		Factory:   f,
-		IOStreams: streams,
-		Gvr:       types.BackupGVR(),
-	}
-	cmd := &cobra.Command{
-		Use:               "describe",
-		Short:             "Describe a backup",
-		ValidArgsFunction: util.ResourceNameCompletionFunc(f, types.BackupGVR()),
-		Example:           describeBackupExample,
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdutil.BehaviorOnFatal(printer.FatalWithRedColor)
-			util.CheckErr(o.Complete(args))
-			util.CheckErr(o.Run())
-		},
-	}
-	return cmd
-}
